+++ package.js
@@ -1,8 +1,8 @@
 Package.describe({
-  name: 'xolvio:inbox-stub',
-  version: '0.1.0',
+  name: 'xolvio:email-stub',
+  version: '0.2.0',
   summary: 'Allows you to inspect sent emails and assert on their content.',
-  git: 'https://github.com/xolvio/meteor-inbox-stub',
+  git: 'https://github.com/xolvio/meteor-email-stub',
   documentation: null,
   debugOnly: true
 });
@@ -10,5 +10,5 @@ Package.describe({
 Package.onUse(function (api) {
   api.versionsFrom('1.0.3.1');
   api.use(['email'], 'server');
-  api.addFiles('inbox.js', 'server');
+  api.addFiles('email.js', 'server');
 });