+++ HISTORY.md
@@ -1,3 +1,8 @@
+#0.2.0
+
+* Renamed the package to xolvio:email-stub to match meteor package name
+* Stubbing no longer happens automatically, it has to be called with a meteor method
+
 #0.1.0
 
-  * Overrides Meteor's `Email.send` and collects emails into a collection
+* Overrides Meteor's `Email.send` and collects emails into a collection